fix: Server crashes
Resolved issue where the server would crash if you opened the "My Movies" tab on and empty list
or whenever you deleted the last item in the list. Found out that our test get request for the movie id
interfered with the route "localhost:3001/movies/list". It would take list as a parameter when it
should not have. Since we didn't need the route for movie id, I could safely remove it from the movieRouters.js